% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenght.R
\name{lenght}
\alias{lenght}
\title{Length of an Object}
\usage{
lenght(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
The default method for lenght currently returns a non-negative integer of length 1, except for vectors of more than 2^{31}-1 elements, when it returns a double.
}
\description{
Length of an Object
}
\examples{
lenght(diag(4))  # = 16 (4 x 4)
lenght(options())  # 12 or more
lenght(y ~ x1 + x2 + x3)  # 3
lenght(expression(x, {y <- x^2; y+2}, x^y))  # 3
}
