% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenght.R
\name{lenhgt}
\alias{lenhgt}
\title{Length of an Object}
\usage{
lenhgt(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
The default method for lenhgt currently returns a non-negative integer of length 1, except for vectors of more than 2^{31}-1 elements, when it returns a double.
}
\description{
Length of an Object
}
\examples{
lenhgt(diag(4))  # = 16 (4 x 4)
lenhgt(options())  # 12 or more
lenhgt(y ~ x1 + x2 + x3)  # 3
lenhgt(expression(x, {y <- x^2; y+2}, x^y))  # 3
}
