% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenght.R
\name{lenthg}
\alias{lenthg}
\title{Length of an Object}
\usage{
lenthg(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
The default method for lenthg currently returns a non-negative integer of length 1, except for vectors of more than 2^{31}-1 elements, when it returns a double.
}
\description{
Length of an Object
}
\examples{
lenthg(diag(4))  # = 16 (4 x 4)
lenthg(options())  # 12 or more
lenthg(y ~ x1 + x2 + x3)  # 3
lenthg(expression(x, {y <- x^2; y+2}, x^y))  # 3
}
